// src/app/modules/operations/model/sale-invoice-detail-ret.ts
export class SaleInvoiceDetailRet {
    ID: number = 0;
    SaleMasterRetID: number = null;
    ItemID: number = null;
    SalePrice: number = null;
    Qty: number = null;
    DiscountPer: number = null;
    DiscountAmount: number = null;
    SalePer: number = null;
    SaleAmount: number = null;
    TotalAmount: number = null;
    NetAmt: number = null;
    ApplyDiscount: boolean = null;
    BuyIn: string = null;
    ApplySaleTax: boolean = null;
}